批量修复审计发现的问题 (P0-A/B/C/D + P1/P2)
P0-A: 17 题 M_PI/M_E 编译错误加 #ifndef 守卫 (968/968 编译通过)
P0-B: WA 归因与修复
  - 057 输入格式(去 T 循环); 058/147/323/196 重写错误算法(int溢出/几何模型错/参数分支占位/三元组定义错)
  - 133 补 p=2,5 反例处理; 059/046 重建可行数据; 189/209 按题意修正(固定答案/越界数据裁剪)
  - 234/235/307/160/200 std 经官方答案锚定无误, 重生成损坏的 .out
P0-C: 探针异常定案: 836(愚人节题)/918(官方答案确为负数)/683/695 免修;
      730/902/919/933/935/951/929/931/369/714/687/689 记入遗留清单(需实现算法)
P0-D: 23 题缺失 .out 全部重生成 (046 数据裁剪至可行范围 K<=2)
P1: 006 __int128 溢出修复 + K=1e5 边界数据; 007 越界 K=0 修复 + 全量重生成
P2: eval.sh python3->python (969 个); 官方答案表 pe_answers.json 入库;
    593/683 内嵌答案回填; audit_full.py 提取正则修正

+++ 058/code/std.cpp
@@ -1,24 +1,24 @@
 #include <bits/stdc++.h>
 using namespace std;
 
-bool is_prime(int n) {
+bool is_prime(long long n) {
     if (n < 2) return false;
     if (n % 2 == 0) return n == 2;
-    for (int i = 3; i * i <= n; i += 2)
+    for (long long i = 3; i * i <= n; i += 2)
         if (n % i == 0) return false;
     return true;
 }
 
 int main() {
     ios::sync_with_stdio(false); cin.tie(nullptr);
-    int P; cin >> P;
-    int total = 1, primes = 0; // center = 1 (not prime)
-    for (int layer = 1; ; layer++) {
-        int side = 2 * layer + 1;
-        int tr = side * side;                  // top-right
-        int tl = tr - (side - 1);              // top-left
-        int bl = tl - (side - 1);              // bottom-left
-        int br = bl - (side - 1);              // bottom-right
+    long long P; cin >> P;
+    long long total = 1, primes = 0; // center = 1 (not prime)
+    for (long long layer = 1; ; layer++) {
+        long long side = 2 * layer + 1;
+        long long tr = side * side;                  // top-right
+        long long tl = tr - (side - 1);              // top-left
+        long long bl = tl - (side - 1);              // bottom-left
+        long long br = bl - (side - 1);              // bottom-right
         if (is_prime(tl)) primes++;
         if (is_prime(tr)) primes++;
         if (is_prime(bl)) primes++;
